requirements.txt
azure-iot-device
python-dotenv
